Remove the TestHelper class that breaks encapsulation.
The TestHelper class was in the adaptor package to use package private
classes and methods in the library.

* Replace Config.setValue with Config.overrideKey.
* Replace WrapperAdaptor.WrapperAdaptorContext with
  UnsupportedAdaptorContext.
